requirements.txt
Django==4.2.7
django-crispy-forms==1.14.0
Pillow
psycopg2
six==1.16.0
whitenoise==6.0.0
django-environ==0.9.0
djangorestframework==3.15.2
django-filter==22.1
drf-spectacular
pre-commit==2.20.0
django-rest-passwordreset==1.3.0
boto3==1.26.66
django-storages==1.13.2
django-cors-headers==4.2.0